// types/habits.ts
export type Area = "Health" | "Knowledge" | "Finance" | "Work" | "Personal";

export type HabitType = "Binary" | "Time" | "Quantity";

export type HabitStatus = "Active" | "Paused" | "Deleted";

export type DayOfWeek =
    | "Monday"
    | "Tuesday"
    | "Wednesday"
    | "Thursday"
    | "Friday"
    | "Saturday"
    | "Sunday";